backend/requirements.txt
# 🚀 GhuntRED-v2 Backend Dependencies
# Django 5.0 with latest optimizations

# Core Django
Django==5.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==23.5

# Database
psycopg2-binary==2.9.9
django-redis==5.4.0

# API Documentation
drf-spectacular==0.27.0

# Authentication & Security
djangorestframework-authtoken==2.0.0

# Task Queue
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
redis==5.0.1

# WebSockets
channels==4.0.0
channels-redis==4.1.0

# File Handling
Pillow==10.1.0

# Environment
python-decouple==3.8
django-environ==0.11.2

# ML Dependencies
numpy==1.24.4
pandas==2.1.4
scikit-learn==1.3.2
joblib==1.3.2

# Text Processing
nltk==3.8.1
spacy==3.7.2

# Deep Learning (Optional)
torch==2.1.2
transformers==4.35.2

# Data Validation
pydantic==2.5.0

# HTTP Requests
requests==2.31.0
httpx==0.25.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3
bleach==6.1.0

# Development
django-debug-toolbar==4.2.0
ipython==8.17.2

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring
sentry-sdk==1.38.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2